requirements.txt
python-telegram-bot
python-dotenv
pytz
requests
persiantools
